Former-commit-id: a460924bbc654027a0cddcc8a322e92ba922b83b

+++ src/parametricStudies/framework_scripts/analyze_results.pl
@@ -27,6 +27,7 @@ use Utilities;
 
 my $tstFile = $ARGV[0];
 my $testNum = $ARGV[1];
+my $uda     = $ARGV[2];
 
 # read XML file into a dom tree
 my $doc = XML::LibXML->load_xml(location => $tstFile);
@@ -83,13 +84,14 @@ if ( ! defined $cmd ){
   die
 }
 
-system("ln -fs $cmd > /dev/null 2>&1");
+if ( ! -l $cmd){
+  system("ln -fs $cmd > /dev/null 2>&1");
+}
 
 #__________________________________
 # Run the comparison tests
 
 my $comp_output = "out.$X.cmp";
-my $uda         = $ups_basename."_$testTitle".".uda";
 
 print "\n\tLaunching: ($cmd_basename -o $comp_output -uda $uda)\n\n";
 `rm -f $comp_output`;
@@ -101,7 +103,7 @@ system("@args")==0 or die("ERROR(analyze_Analysis.pl):\tFailed running: (@args))
 my $L2norm = `cat $comp_output`;
 chomp($L2norm);
 `echo $X $L2norm >> L2norm.dat`;
-#system("rm $comp_output");
+`rm -f $comp_output`;
   
 
 #______________________________________________________________________
@@ -115,10 +117,12 @@ if ( length $gpData ) {
   my $title = $doc->findvalue( '/start/gnuplot/title' );
   my $xlabel= $doc->findvalue( '/start/gnuplot/xlabel' );
   my $ylabel= $doc->findvalue( '/start/gnuplot/ylabel' );
+  my $label = $doc->findvalue( '/start/gnuplot/label' );
 
-  system("sed", "-i", "s/#title/set title \"$title\"/g", "$gpFile");
+  system("sed", "-i", "s/#title/set title   \"$title\"/g",  "$gpFile");
   system("sed", "-i", "s/#xlabel/set xlabel \"$xlabel\"/g", "$gpFile");
   system("sed", "-i", "s/#ylabel/set ylabel \"$ylabel\"/g", "$gpFile");
+  system("sed", "-i", "s/#label/set label     $label/g",      "$gpFile");
 
   print "Now plotting using the modified gnuplot script ($gpFile) \n";
   system("gnuplot $gpFile > gp.out 2>&1");
